Core: Extract processing logic into ProcessingQueue
